Add resume to UpdateMe mutation
